feat(drizzle): solder les dettes de DDL avant que la migration 0000 soit gravée
Le premier fichier de migration publié devient immuable : son empreinte est
vérifiée à chaque exécution, et toute erreur de DDL qu'il contient ne se corrige
plus que par une migration correctrice, visible à jamais dans l'historique de
chaque utilisateur. Les défauts ci-dessous coûtent zéro aujourd'hui — ils
coûteraient une migration publique après #17.

Deux capacités neuves au colKit :

- `int64` — `int` est un entier 32 bits SIGNÉ sur les trois dialectes. Un
  compteur qui le dépasse ne rend pas d'erreur lisible : il tronque ou refuse
  l'écriture, des années plus tard. `int` reste pour ce que le domaine borne
  (nombre de chiffres, code de statut).
- `enum` — texte borné par une contrainte `CHECK` émise dans le `CREATE TABLE`,
  en développement comme en production, depuis la MÊME déclaration.

Trois colonnes corrigées :

- `webauthn_credential.signCount` → `int64` : WebAuthn définit ce compteur
  anti-clonage en uint32 (4 294 967 295), au-delà de l'entier signé, et la
  valeur vient de l'authenticator — rien ne la borne côté serveur.
- `idempotency_key.state` → `enum ["if","done"]` : l'union ne vivait que sur le
  type `Row`, donc elle ne protégeait que le code passant par le store. Une
  commande d'exploitation pouvait poser un troisième état, que la reprise après
  incident aurait lu comme une réservation ni vivante ni terminée.
- `totp_secret.lastUsedStep` → `int64` (défaut TROUVÉ en séance, pas au
  cahier) : T = floor(epochSeconds / period), et `period` n'a pas de borne basse
  (schéma Zod `positive()`). À period = 1 la tranche vaut l'horodatage Unix, qui
  dépasse l'entier 32 bits en 2038 — l'anti-rejeu cesserait de retenir, en
  silence.

Un `CHECK` se grave à vie : il ne se pose que sur un ensemble fermé par
construction. Les catalogues appelés à s'étendre (algorithme de hachage,
catégorie d'audit, type de sujet) restent en `text`, leur union sur le `Row`.

Preuves — les cinq contrôles neufs ont été VUS ROUGES par débranchement (retrait
de l'émission dans le DDL, retrait de la fabrique de contraintes, `int64` remis
en `integer` PG) : DDL réellement exécuté relu au catalogue, insertion ET mise à
jour hors énumération refusées par le serveur, contrainte présente et mordante
sur PostgreSQL, MariaDB 11.4 et MySQL Community 8.4.

drizzle 1030 verts / 3 dialectes (gates PG + MySQL exercés), security 1034 verts.

+++ src/packages/@nodefony/drizzle/nodefony/entity/totpSecretEntity.ts
@@ -50,8 +50,15 @@ const TOTP_SECRET_TABLE_SPEC = {
     recoveryCodes: { kind: "json", notNull: true, defaultFn: () => [] },
     /** Confirmation de l'enrôlement (epoch ms) ou `null` (anti-lock-out). */
     confirmedAt: { kind: "epochMs" },
-    /** Dernière tranche `T` validée (RFC 6238 §5.2, anti-rejeu) ou `null`. */
-    lastUsedStep: { kind: "int" },
+    /**
+     * Dernière tranche `T` validée (RFC 6238 §5.2, anti-rejeu) ou `null`.
+     *
+     * `int64` : `T = floor(epochSeconds / period)`, et `period` n'a pas de
+     * borne basse (schéma Zod : `positive()`). À `period` = 1 la tranche vaut
+     * l'horodatage Unix lui-même, qui dépasse l'entier 32 bits signé en 2038 —
+     * l'anti-rejeu cesserait alors de retenir, en silence.
+     */
+    lastUsedStep: { kind: "int64" },
     /** Création (epoch ms). */
     createdAt: { kind: "epochMs", notNull: true },
     /** Dernier usage réussi (epoch ms) ou `null`. */
